pyproject.toml
[project]
name = "textsearch"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "kiwipiepy>=0.18.0",
    "python-mecab-ko>=1.0.0",
    "konlpy>=0.6.0",
    "elasticsearch>=8.0.0",
    "qdrant-client>=1.7.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "pdfplumber>=0.10.0",
    "anthropic>=0.20.0",
]
